+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.6.2 (Unreleased)
 
 - Fixed rolled log files created with trailing sequence sometimes created on app startup ([#136](https://github.com/tetsuo13/TeamsStatusPub/issues/136))
+- Prevent multiple About windows ([#129](https://github.com/tetsuo13/TeamsStatusPub/issues/129))
 
 ## 1.6.1 (2024-08-28)